use separated query clause for query object
